refactor: changed resource files' names to match Kotlin conventions
