+++ compiler-plugin-test/src/test/resources/kotlin/OutsideAggregateForExampleAggregate.kt
