+++ compiler-plugin-test/src/test/resources/kotlin/OutsideAggregateForNeighboring.kt
