+++ compiler-plugin-test/src/test/resources/kotlin/OutsideAggregateListOfForEachExampleAggregate.kt
