+++ compiler-plugin-test/src/test/resources/kotlin/OutsideAggregateListOfForEachNeighboring.kt
