+++ compiler-plugin-test/src/test/resources/kotlin/SimpleItListOfForEachNeighboring.kt
@@ -1,4 +1,5 @@
 import it.unibo.collektive.aggregate.api.Aggregate
+import it.unibo.collektive.Collektive.Companion.aggregate
 
 fun Aggregate<Int>.exampleAggregate() {}
 
@@ -8,8 +9,6 @@ fun Aggregate<Int>.exampleAggregate() {}
 
 fun Aggregate<Int>.entry() {
     listOf(1,2,3).forEach {
-        alignedOn(0) {
-            exampleAggregate()
-        }
+        neighboring(0)
     }
 }
